Update client-libraries.md
We recently release an early version of https://github.com/wp-headless/fetch that is now being used in production. 

It's not an easy decision to fragment the ecosystem when there are other offerings. We believe our approach is significantly different enough to warrant this risk. We did so for the following reasons:

* Long unresolved [browser issues](https://github.com/WP-API/node-wpapi/issues/438)
* Bloated [packages size](https://bundlephobia.com/result?p=wpapi@0.12.1)
* No [tree-shakable](https://webpack.js.org/guides/tree-shaking/) ESM or CJS build available
* Opinionated API that attempts to do more then is needed.
* Lack of automated browser testing and coverage
* + more

wp-headless/fetch will be a valuable option to the community and therefore should be listed. Thanks! :-)