Update controller-classes.md
Removed $request param from get_item_schema() since this function does not actually support parameters when used in the WP_REST_Posts_Controller class. Including the param here is misleading since it is not needed, even in this custom controller. If the reader uses this example with the WP_REST_Posts_Controller class, an error will occur.